+++ src/__mocks__/axios.ts
@@ -1,4 +1,0 @@
-export default {
-    create: jest.fn(),
-    get: jest.fn(),
-};