+++ src/__mocks__/fs.ts
@@ -1,2 +1,0 @@
-const fs = jest.genMockFromModule<any>('fs');
-module.exports = fs;