+++ src/index.ts
@@ -1,4 +1,2 @@
-export { PublicESIService } from './public-esi.service';
-export { CacheController } from './cache.controller';
 export { EVE } from './eve.helper';
 export * from './interface.helper';