+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the eve-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### BREAKING CHANGES
+- Moved PublicEsiService and CacheController to esi-service package.
+
 ## [1.0.1] 2019-04-26
 ### Fixed
 - Some data interfaces not being arrays of their units.